Modification headers leftmost_bits

+++ bits_leftmost/student/student_code.h
@@ -1,1 +1,2 @@
-int get_3_leftmost_bits(uint32_t x);
+#include <stdint.h>
+uint8_t get_3_leftmost_bits(uint32_t x);
